BLADE-533 fix tests

+++ cli/src/main/java/com/liferay/blade/cli/Extensions.java
@@ -205,7 +205,9 @@ public class Extensions implements Closeable {
 
 	@Override
 	public void close() throws IOException {
-		FileUtil.deleteDir(_embeddedTemplatesPath);
+		if ((_embeddedTemplatesPath != null) && Files.exists(_embeddedTemplatesPath)) {
+			FileUtil.deleteDir(_embeddedTemplatesPath);
+		}
 	}
 
 	public Map<String, BaseCommand<? extends BaseArgs>> getCommands() throws Exception {
